Seperate the useItems core from the useListItems hook
